requirements.txt
# Smart Resume and Job Matcher - Dependencies
# Core Framework
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0

# Document Parsing
pypdf2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0

# Embeddings
sentence-transformers>=2.2.0
ollama>=0.1.0

# Vector Operations & Similarity
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4

# Data Processing
pydantic>=2.0.0
pandas>=2.0.0

# Visualization
matplotlib

# Notebook Support
jupyter>=1.0.0
ipykernel>=6.0.0

# Google AI SDK (optional - for Gemini/Vertex AI)
google-generativeai>=0.3.0
